backend/requirements.txt
Flask==2.2.3
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
geopy==2.3.0
numpy==1.23.5
opencage==2.1.0
Pillow==9.3.0
PyJWT==2.6.0
python-dotenv==1.0.0
python_bcrypt==0.3.2
Requests==2.28.2
SQLAlchemy==2.0.9
tensorflow==2.12.0
waitress==2.1.2